Anti-symmetric kernel SVMkernel matrix and manifold, kernel classes overhaul, other minor fixes and improvements.

+++ SVRRoot/SVRBusiness/src/dq_scaling_factors_service_impl.cu
@@ -34,7 +34,7 @@ void calc_mean_max(
     double &max)
 {
     mean = thrust::reduce(thrust::device, start_it, end_it);
-    mean /= double(rows);
+    mean /= rows;
 
     thrust::transform(thrust::device, start_it, end_it, start_it, abs_functor());
     thrust::sort(thrust::device, start_it, end_it);
